requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sqlalchemy>=2.0.0
streamlit>=1.28.0
joblib>=1.3.0
requests>=2.31.0
plotly>=5.18.0
python-dotenv>=1.0.0
